refact: unify pix and installment options
