+++ src/pages/paymentMethod/index.tsx
@@ -2,10 +2,11 @@ import LogoContainer from "../../components/logoContainer";
 import { getData } from "../../fakeAPI/getData";
 import PixOption from "../../components/pixOption";
 import InstallmentOption from "../../components/installmentOption";
+import PaymentOption from "../../components/paymentOption";
 import {
-  OptionsContainer,
   PageContainer,
   PaymentText,
+  OptionsContainer,
   OptionChip,
 } from "./style";
 import React, { useState } from "react";
@@ -28,31 +29,37 @@ export default function PaymentMethod() {
       <LogoContainer />
       <PaymentText>{`${userInfo.name}, como você gostaria de pagar?`}</PaymentText>
       <OptionsContainer>
-        <OptionChip variant="filled" label="Pix" size="small" />
-        <PixOption
+        <OptionChip label="Pix" />
+        <PaymentOption
           selectedValue={selectedValue}
           installments={pixOption.installments}
           installmentValue={pixOption.installmentValue}
-          cashbackAmount={pixOption.cashbackAmount}
-          cashbackPercentage={pixOption.cashbackPercentage}
           handleRadioChange={handleRadioChange}
           handleButtonClick={handleButtonClick}
-        />
+        >
+          <PixOption
+            cashbackAmount={pixOption.cashbackAmount}
+            cashbackPercentage={pixOption.cashbackPercentage}
+          />
+        </PaymentOption>
       </OptionsContainer>
       <OptionsContainer>
-        <OptionChip variant="filled" label="Pix Parcelado" size="small" />
+        <OptionChip label="Pix Parcelado" />
         {installmentOptions.map((installment) => {
           return (
-            <InstallmentOption
+            <PaymentOption
+              handleButtonClick={handleButtonClick}
+              handleRadioChange={handleRadioChange}
+              installmentValue={installment.installmentValue}
               selectedValue={selectedValue}
               installments={installment.installments}
-              installmentValue={installment.installmentValue}
-              totalAmount={installment.totalAmount}
-              discount={installment.discount}
-              key={installment.installments}
-              handleRadioChange={handleRadioChange}
-              handleButtonClick={handleButtonClick}
-            />
+            >
+              <InstallmentOption
+                totalAmount={installment.totalAmount}
+                discount={installment.discount}
+                key={installment.installments}
+              />
+            </PaymentOption>
           );
         })}
       </OptionsContainer>
